apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
anthropic==0.40.0
google-api-python-client==2.155.0
google-auth==2.36.0
dnspython>=2.6
aiosmtplib>=3.0
